pyproject.toml
[project]
name = "poooli"
version = "0.1.0"
description = "A python library for printing on the Poooli thermal printer."
authors = [
    {name = "Christoph Minixhofer", email = "christoph.minixhofer@gmail.com"},
]
dependencies = [
    "Pillow>=9.3.0",
    "pybluez>=0.23",
]
requires-python = ">=3.8"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"

[tool.pdm]
